% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.cos.R
\name{adj.cos}
\alias{adj.cos}
\title{Cosine adjustment term, not the series.}
\usage{
adj.cos(distance, scaling, adj.order)
}
\arguments{
\item{distance}{perpendicular distance vector/scalar}

\item{scaling}{scale parameter}

\item{adj.order}{the adjustment order}
}
\value{
scalar or vector containing the cosine adjustment term for every
value in \code{distance} argument
}
\description{
For internal use only -- not to be called by 'mrds' or 'Distance' users 
directly.
}
\author{
Felix Petersma
}
