% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.series.grad.herm.R
\name{adj.series.grad.herm}
\alias{adj.series.grad.herm}
\title{Series of the gradient of the Hermite polynomial adjustment series w.r.t. the 
scaled distance.}
\usage{
adj.series.grad.herm(
  distance,
  scaling = 1,
  adj.order,
  adj.parm = NULL,
  adj.exp = FALSE
)
}
\arguments{
\item{distance}{perpendicular distance vector/scalar}

\item{scaling}{scale parameter}

\item{adj.order}{the adjustment order}

\item{adj.parm}{vector of parameters (a_j)}

\item{adj.exp}{boolean, defaults to FALSE}
}
\value{
scalar or vector containing the gradient of the Hermite adjustment 
series for every value in \code{distance} argument
}
\description{
For internal use only -- not to be called by 'mrds' or 'Distance' users 
directly.
}
\author{
Felix Petersma
}
