% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R
\name{AIC.ddf}
\alias{AIC.ddf}
\alias{AIC.ds}
\alias{AIC.io}
\alias{AIC.io.fi}
\alias{AIC.rem}
\alias{AIC.rem.fi}
\alias{AIC.trial}
\alias{AIC.trial.fi}
\title{Akaike's An Information Criterion for detection functions}
\usage{
\method{AIC}{ddf}(object, ..., k = 2)
}
\arguments{
\item{object}{a fitted detection function object}

\item{\dots}{optionally more fitted model objects.}

\item{k}{penalty per parameter to be used; the default \code{k = 2} is the
"classical" AIC}
}
\description{
Extract the AIC from a fitted detection function.
}
\author{
David L Miller
}
