% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.line.cond.R
\name{average.line.cond}
\alias{average.line.cond}
\title{Average conditional detection function line for plotting}
\usage{
average.line.cond(finebr, obs, model)
}
\arguments{
\item{finebr}{set of fine breaks in distance over which detection function
values are averaged and plotted}

\item{obs}{value of observer for averaging (1-2 individual observers)}

\item{model}{ddf model object}
}
\value{
list with 2 elements:
  \tabular{ll}{\code{xgrid} \tab vector of gridded distance values \cr
  \code{values} \tab vector of average detection function values at
   the \code{xgrid} values\cr}
}
\description{
For models with covariates the detection probability for each observation
can vary.  This function computes an average value for a set of distances to
plot an average line to graphically represent the fitted model in plots that
compare histograms and the scatter of individual estimated detection
probabilities.
}
\note{
Internal function called from plot functions for \code{ddf} objects
}
\author{
Jeff Laake
}
\keyword{utility}
