\name{bears}
\alias{bears}
\title{Grizzly bear aerial survey observations...}

\description{
  Grizzly bear aerial survey observations Observation data
  from fixed wing aerial survey of grizzly bears in Alaska.
  Independent dual observers were used with different
  viewing through flat windows. Good example of the gamma
  detection function.
}

\details{
  These data are in the standard format for analysis with
  mrds.
}
\docType{data}
\references{Becker, E. F. and P. X. Quang. 2009. A gamma-shaped detection
function for line transect surveys with mark-recapture and covariate data.
Journal of Agricultural Biological and Environmental Statistics
14:207-223.}
\keyword{datasets}
