\name{check.bounds}
\alias{check.bounds}
\title{Check bounds during optimsations}
\usage{
check.bounds(lt, lowerbounds, upperbounds, ddfobj, showit, setlower, setupper)
}
\description{
Simple internal function to check that the optimisation
didn't hit bounds. Based on code that used to live in
\code{detfct.fit.opt}.
}
\author{
Dave Miller; Jeff Laake
}

