% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Nht.R
\name{compute.Nht}
\alias{compute.Nht}
\title{Horvitz-Thompson estimates 1/p_i or s_i/p_i}
\usage{
compute.Nht(pdot, group = TRUE, size = NULL)
}
\arguments{
\item{pdot}{vector of estimated detection probabilities}

\item{group}{if TRUE (do group abundance); FALSE(do individual abundance)}

\item{size}{vector of group size values for clustered populations}
}
\value{
vector of H-T components for abundance estimate
}
\description{
Compute individual components of Horvitz-Thompson abundance estimate in
covered region for a particular subset of the data depending on value of
group = TRUE (do group abundance); FALSE(do individual abundance)
}
\note{
Internal function called by \code{\link{covered.region.dht}}
}
\author{
Jeff Laake
}
\keyword{utility}
