% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.bins.R
\name{create.bins}
\alias{create.bins}
\title{Create bins from a set of binned distances and a set of cutpoints.}
\usage{
create.bins(data, cutpoints)
}
\arguments{
\item{data}{`data.frame` with at least the column `distance`.}

\item{cutpoints}{vector of cutpoints for the bins}
}
\value{
argument `data` with two extra columns `distbegin` and
       `distend`.
}
\description{
This is an internal routine and shouldn't be necessary in normal analyses.
}
\author{
David L. Miller
}
