% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcds_tools.R
\name{create.command.file}
\alias{create.command.file}
\title{create.command.file}
\usage{
create.command.file(dsmodel = call(), data, method, meta.data, control)
}
\arguments{
\item{dsmodel}{distance sampling model specification}

\item{data}{dataframe containing data to be analyzed}

\item{method}{analysis method}

\item{meta.data}{list containing settings controlling data structure}

\item{control}{list containing settings controlling model fitting}
}
\description{
create.command.file
}
\author{
Jonah McArthur
}
