% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_document}
\alias{create_document}
\alias{create_knitr}
\title{Create Rmarkdown document using knitr}
\usage{
create_document(type="html")
       create_knitr(output,type="html")
}
\arguments{
\item{output}{result list from shinyApp}

\item{type}{"html","pdf" or "word"}
}
\value{
filename of result from render in markdown
}
\description{
Create Rmarkdown document using knitr
}
\author{
Jeff Laake
}

