% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcds_tools.R
\name{download_MCDS_dot_exe}
\alias{download_MCDS_dot_exe}
\title{Install the MCDS.exe binary from the web}
\usage{
download_MCDS_dot_exe()
}
\value{
Nothing, just downloads and installs the binary.
}
\description{
This function will download the `MCDS.exe` binary from the distance sampling
website and install it.
}
\details{
If you already have a copy of `MCDS.exe` on your computer you can move it to
the correct directory manually. You can find where this is by running the
following: `system.file(package="mrds")`.
}
