% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ds_cov_levels.R
\name{ds_cov_levels}
\alias{ds_cov_levels}
\title{Plot covariate levels for a detection function}
\usage{
ds_cov_levels(x, distances)
}
\description{
Utility routine to plot the levels of a covariate as lines on a plot. This only works for scale function covariates at the moment.
}
\author{
David L Miller
}

