\name{errors}
\alias{errors}
\title{Error function...}
\usage{errors(errmsg, errmode="report")}

\description{
  Error function Writes error messages for various errors
  that can occur
}
\value{None}
\author{Dave Miller}
\arguments{
  \item{errmsg}{the message to be stored/printed (optional)}
  \item{errmode}{report or print errors (default report)}
}
