\name{gof.io.fi}
\alias{gof.io.fi}
\title{Compute chi-square gof test for io.fi models}
\usage{
gof.io.fi(model, breaks = NULL, nc = NULL)
}
\arguments{
  \item{model}{ddf model object}

  \item{breaks}{distance cut points}

  \item{nc}{number of distance classes}
}
\value{
list with chi-square value, df and p-value
}
\description{
Compute chi-square gof test for io.fi models
}

