\name{gof.trial.fi}
\alias{gof.trial.fi}
\title{Compute chi-square gof test for trial.fi models}
\usage{
gof.trial.fi(model, breaks = NULL, nc = NULL)
}
\arguments{
  \item{model}{ddf model object}

  \item{breaks}{distance cut points}

  \item{nc}{number of distance classes}
}
\value{
lists with observed,expected, chi-square value, df and
p-value
}
\description{
Compute chi-square gof test for trial.fi models
}

