\name{hermite.poly}
\alias{hermite.poly}
\title{Hermite polynomial generating function}
\usage{
hermite.poly(z, n)
}
\arguments{
  \item{z}{points to evaluate}

  \item{n}{order Hermite polynomial

  Using formula from
  http://functions.wolfram.com/Polynomials/HermiteH/02/0001/
  This may not be the fastest way of finding the value.}
}
\description{
Hermite polynomial generating function
}

