\name{keyfct.gamma}
\alias{keyfct.gamma}
\title{Gamma key function}
\usage{
keyfct.gamma(distance, key.scale, key.shape)
}
\arguments{
  \item{distance}{perpendicular distance vector}

  \item{key.scale}{vector of scale values}

  \item{key.shape}{vector of shape values}
}
\value{
vector of probabilities
}
\description{
Gamma key function
}

