% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyfct.grad.hn.R
\name{keyfct.grad.hn}
\alias{keyfct.grad.hn}
\title{The gradient of the half-normal key function}
\usage{
keyfct.grad.hn(distance, key.scale)
}
\arguments{
\item{distance}{perpendicular distance vector}

\item{key.scale}{vector of scale values}
}
\value{
vector of derivatives of the half-normal key function w.r.t. the
scale parameter
}
\description{
The key function contains one parameter, the scale. 
Current implementation assumes that scaled dist is x/scale, not x/width
}
\details{
d key / d scale = exp(-y ^ 2 / (2 scale ^ 2)) * (y ^ 2 / scale ^ 3)
}
