% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyfct.th.R
\name{keyfct.th1}
\alias{keyfct.th1}
\title{Threshold key function}
\usage{
keyfct.th1(distance, key.scale, key.shape)
}
\arguments{
\item{distance}{perpendicular distance vector}

\item{key.scale}{vector of scale values}

\item{key.shape}{vector of shape values}
}
\value{
vector of probabilities
}
\description{
Threshold key function
}
