% Generated by roxygen2 (4.0.1): do not edit by hand
\name{keyfct.th2}
\alias{keyfct.th2}
\title{Threshold key function}
\usage{
keyfct.th2(distance, key.scale, key.shape)
}
\arguments{
\item{distance}{perpendicular distance vector}

\item{key.scale}{vector of scale values}

\item{key.shape}{vector of shape values}
}
\value{
vector of probabilities
}
\description{
Threshold key function
}

