% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticbyx.R
\name{logisticbyx}
\alias{logisticbyx}
\title{Logistic as a function of covariates}
\usage{
logisticbyx(distance, x, models, beta, point)
}
\arguments{
\item{distance}{vector of distance values}

\item{x}{covariate data}

\item{models}{model list}

\item{beta}{logistic parameters}

\item{point}{\code{TRUE} if a point transect model}
}
\value{
vector of probabilities
}
\description{
treats logistic as a function of covariates; for a given covariate
combination it computes function at with those covariate values at a
range of distances
}
\author{
Jeff Laake
}
