% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticdupbyx_fast.R
\name{logisticdupbyx_fast}
\alias{logisticdupbyx_fast}
\title{Logistic for duplicates as a function of covariates (fast)}
\usage{
logisticdupbyx_fast(distance, x1, x2, models, beta, point, beta_distance)
}
\arguments{
\item{distance}{vector of distance values}

\item{x1}{linear predictor for 1, without distance}

\item{x2}{linear predictor for 2, without distance}

\item{models}{model list}

\item{beta}{logistic parameters}

\item{point}{\code{TRUE} for point transect data}

\item{beta_distance}{parameter for distance}
}
\description{
As \code{\link{logisticdupbyx}}, but faster when distance is a covariate (but no interactions with distance occur.
}
\author{
David L Miller
}
