% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(p)
}
\arguments{
\item{p}{probability}
}
\value{
\code{ logit(p)} returns [log(p/(1-p)]
}
\description{
Computes logit transformation.
}
\author{
Jeff Laake
}
\keyword{utility}
