% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.ddf}
\alias{logLik.ddf}
\alias{logLik.ds}
\alias{logLik.io}
\alias{logLik.io.fi}
\alias{logLik.rem}
\alias{logLik.rem.fi}
\alias{logLik.trial}
\alias{logLik.trial.fi}
\title{log-likelihood value for a fitted detection function}
\usage{
\method{logLik}{ddf}(object, ...)
}
\arguments{
\item{object}{a fitted detection function model object}

\item{\dots}{included for S3 completeness, but ignored}
}
\value{
a numeric value giving the log-likelihood with two attributes:
\code{"df"} the "degrees of freedom" for the model (number of parameters)
and \code{"nobs"} the number of observations used to fit the model
}
\description{
Extract the log-likelihood from a fitted detection function.
}
\author{
David L Miller
}
