% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcds_tools.R
\name{create_command_file}
\alias{create_command_file}
\alias{MCDS}
\title{Run MCDS.exe as a backend for mrds}
\usage{
create_command_file(
  dsmodel = call(),
  mrmodel = call(),
  data,
  method,
  meta.data,
  control
)
}
\description{
Rather than use the R code provided in `mrds`, one can also use the binary
of MCDS.exe, to reproduce the results given by Distance for Windows. There is
no guarantee that one approach is "better" than the other, but `mrds` will
select the model with the better likelihood and provide answers to this.
}
\section{Running MCDS.exe on non-Windows platforms}{

One can still use MCDS.exe even if you are not running a Windows computer. To
do this one will need to install `wine` a Windows emulator. It is important
to get a copy of `wine` which can run 32-bit programs. On macOS, this can be
achieved using the `homebrew` package management system and installing the
`wine-stable` package. You may need to change the `control$winebin` to be
`wine` or `wine64`, depending on your system's setup.
}

\author{
David L Miller and Joanna McArthur
}
