% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrds-package.R
\name{mrds-package}
\alias{mrds-package}
\alias{mrds}
\title{Mark-Recapture Distance Sampling (mrds)}
\description{
This package implements both conventional distance sampling 
analysis and mark-recapture distance sampling methods for when
detection on the transect is not certain. Mark-recapture distance sampling methods are described in 
\insertCite{borchers1988;textual}{mrds}, \insertCite{laake1999;textual}{mrds} 
and \insertCite{borchers2006;textual}{mrds}
}
\details{
Examples of distance sampling analyses are available at
\url{https://distancesampling.org/resources/vignettes.html}.

For help with distance sampling and this package, there is a Google Group
\url{https://groups.google.com/forum/#!forum/distance-sampling}.
}
\references{
\insertAllCited{}
}
\author{
Jeff Laake <jeff.laake@noaa.gov>,
        David Borchers <dlb@mcs.st-and.ac.uk>,
        Len Thomas <len@mcs.st-and.ac.uk>,
        David L. Miller <dave@ninepointeightone.net>,
        Jon Bishop <jonb@mcs.st-and.ac.uk>,
        Felix Petersma <ftp@st-andrews.ac.uk>
}
\keyword{"_PACKAGE"}
\keyword{package}
