% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCovered.R
\name{NCovered}
\alias{NCovered}
\alias{NCovered.ds}
\alias{NCovered.io}
\alias{NCovered.io.fi}
\alias{NCovered.trial}
\alias{NCovered.trial.fi}
\alias{NCovered.rem}
\alias{NCovered.rem.fi}
\title{Compute estimated abundance in covered (sampled) region}
\usage{
NCovered(par, model = NULL, group = TRUE)
}
\arguments{
\item{par}{parameter values (used when computing derivatives wrt parameter
uncertainty); if NULL parameter values in \code{model} are used}

\item{model}{ddf model object}

\item{group}{if TRUE computes group abundance and if FALSE individual
abundance}
}
\value{
abundance estimate
}
\description{
Generic function that computes abundance within the covered region.  It
calls method (class) specific functions for the computation.
}
\author{
Jeff Laake
}
\keyword{utility}
