% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlminb_wrapper.R
\name{nlminb_wrapper}
\alias{nlminb_wrapper}
\title{Wrapper around \code{nlminb}}
\usage{
nlminb_wrapper(par, ll, ugr = NULL, lower = NULL, upper = NULL, mcontrol,
  hess = NULL, ddfobj, data, ...)
}
\arguments{
\item{par}{starting parameters}

\item{ll}{log likelihood function}

\item{ugr}{gradent function}

\item{lower}{lower bounds on parameters}

\item{upper}{upper bounds on parameters}

\item{mcontrol}{control options}

\item{hess}{hessian function}

\item{ddfobj}{detection function specification object}

\item{data}{the data}

\item{\dots}{anything else to pass to \code{ll}}
}
\value{
\code{optimx} object
}
\description{
This is a wrapper around nlminb to use scaling, as this is not available (nor will it be) in \code{\link{optimx}}.
}
\author{
David L Miller, modified from \code{optimx.run} by JC Nash, R Varadhan, G Grothendieck.
}
