\name{p.det}
\alias{p.det}
\title{Double-platform detection probability...}
\usage{p.det(dpformula, dplink, dppars, dpdata)}

\description{
  Double-platform detection probability Computes detection
  probability for detection function computed from
  mark-recapture data with possibly different link
  functions.
}
\value{vector of predicted detection probabilities}
\author{?????}
\arguments{
  \item{dpformula}{formula for detection function}
  \item{dplink}{link function ("logit","loglog","cloglog"}
  \item{dppars}{parameter vector}
  \item{dpdata}{double platform data}
}
