% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_dist_table.R
\name{p_dist_table}
\alias{p_dist_table}
\title{Distribution of probabilities of detection}
\usage{
p_dist_table(object, bins = seq(0, 1, by = 0.1), proportion = FALSE)
}
\arguments{
\item{object}{fitted detection function}

\item{bins}{how the results should be binned}

\item{proportion}{should proportions be returned as well as counts?}
}
\value{
a \code{data.frame} with probability bins, counts and (optionally) proportions. The object has an attribute \code{p_range} which contains the range of estimated detection probabilities
}
\description{
Generate a table of frequencies of probability of detection from a detection function model. This is particularly useful when employing covariates, as it can indicate if there are detections with very small \hat{P}s that can be unduly influential when calculating abundance estimates.
}
\author{
David L Miller
}
