% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.optimx.R
\name{parse.optimx}
\alias{parse.optimx}
\title{Parse optimx results and present a nice object}
\usage{
parse.optimx(lt, lnl.last, par.last)
}
\arguments{
\item{lt}{an optimx object}

\item{lnl.last}{last value of the log likelihood}

\item{par.last}{last value of the parameters}
}
\value{
\code{lt} object that can be used later on
}
\description{
Take the resulting object from a call to optimx and make it into an object that mrds wants to talk to.
}
