% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_avg_df.R
\name{plot_avg_df}
\alias{plot_avg_df}
\title{Calculate values of the average detection function when covariates are present}
\usage{
plot_avg_df(xgrid, width, xmat, ddfobj, selected, int.range, pdot, normalize,
  range.varies)
}
\description{
Calculate values of the average detection function when covariates are present
}

