% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ddf.gof.R
\name{print.ddf.gof}
\alias{print.ddf.gof}
\title{Prints results of goodness of fit tests for detection functions}
\usage{
\method{print}{ddf.gof}(x, ...)
}
\arguments{
\item{x}{result of call to \code{\link{ddf.gof}}}

\item{\dots}{unused unspecified arguments for generic print}
}
\value{
None
}
\description{
Provides formatted output for results of goodness of fit tests: chi-square,
Kolmogorv-Smirnov and Cramer-von Mises test as approrpriate.
}
\seealso{
\code{\link{ddf.gof}}
}
\author{
Jeff Laake
}
\keyword{utility}
