\name{print.ddf}
\alias{print.ddf}
\title{Simple pretty printer for distance sampling analyses}
\usage{
  \method{print}{ddf} (x, ...)
}
\arguments{
  \item{x}{a \code{ddf} object}

  \item{...}{not passed through, just for S3
  compatibility.}
}
\description{
  Simply prints out a brief description of the model which
  was fitted. For more detailed information use
  \code{\link{summary}}.
}
\author{
  David L. Miller
}

