% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dht.R
\name{print.dht}
\alias{print.dht}
\title{Prints density and abundance estimates}
\usage{
\method{print}{dht}(x, cor = FALSE, bysample = FALSE, vcmatrices = FALSE,
  ...)
}
\arguments{
\item{x}{dht object that results from call to dht for a specific ddf object}

\item{cor}{if TRUE outputs correlation matrix of estimates}

\item{bysample}{if TRUE, prints results for each sample}

\item{vcmatrices}{if TRUE, prints variance-covariance matrices}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\value{
None
}
\description{
Outputs summary statistics, abundance and density by region (if any) and optionally a correlation matrix if more than one region.
}
\seealso{
\code{\link{dht}}
}
\author{
Jeff Laake
}
\keyword{utility}
