% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.dist.table.R
\name{print.p_dist_table}
\alias{print.p_dist_table}
\title{Print distribution of probabilities of detection}
\usage{
\method{print}{p_dist_table}(x, digits = 2, ...)
}
\arguments{
\item{x}{output from \code{\link{p_dist_table}}}

\item{digits}{number of significant digits to print}

\item{\dots}{other arguments to be passed to \code{\link{print.data.frame}}}
}
\value{
just prints the table and the range of ps
}
\description{
Just a pretty printer for the table of probabilities of detection.
}
\author{
David L Miller
}
