\name{print.summary.io}
\title{Print summary of distance detection function model object...}
\usage{\method{print}{summary.io} (x, ...)}

\description{
  Print summary of distance detection function model object
  :provides a brief summary of data and fitted detection
  probability model parameters, model selection criterion,
  and optionally abundance in the covered (sampled) region
  and its standard error. What is printed depends on the
  corresponding call to summary.
}
\alias{print.summary.io}
\value{NULL}
\author{Jeff Laake}
\seealso{\code{\link{summary.io}}}
\keyword{utility}
\arguments{
  \item{x}{a summary of \code{ddf} model object}
  \item{\dots}{unspecified and unused arguments for S3 consistency}
}
