% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.se.R
\name{prob.se}
\alias{prob.se}
\title{Average p and average p(0) variance}
\usage{
prob.se(model, fct, vcov, observer = NULL, fittedmodel = NULL)
}
\arguments{
\item{model}{ddf model object}

\item{fct}{function of detection probabilities; currently only
average (over covariates) detection probability p integrated over distance
or average (over covariates) detection probability at distance 0; p(0)}

\item{vcov}{variance-covariance matrix of parameter estimates}

\item{observer}{1,2,3 for primary, secondary, or duplicates for average
p(0); passed to fct}

\item{fittedmodel}{full fitted ddf model when \code{trial.fi} or
\code{io.fi} is called from \code{trial} or \code{io} respectively}
}
\value{
\item{var}{variance} \item{partial}{partial derivatives of
  parameters with respect to fct} \item{covar}{covariance of n and average p
  or p(0)}
}
\description{
Computes components of variance for average p=n/N and average p(0) with
weights based on empirical covariate distribution, if it contains covariates.
}
\details{
Need to add equations here as I do not think they exist in any of the texts.
These should probably be checked with simulation.
}
\seealso{
prob.deriv
}
\author{
Jeff Laake
}
