\name{pronghorn}
\alias{pronghorn}
\title{Pronghorn aerial survey data from Wyoming...}

\description{
  Pronghorn aerial survey data from Wyoming Detections of
  pronghorn from fixed-wing aerial surveys in Southeastern
  Wyoming using four angular bins defined by strut marks.
  Illustrates data where altitude above ground level (AGL)
  varies during the survey.
}

\details{
  Each record is an observed cluster of pronghorn.  The
  data provide the stratum for the observation, the
  direction of travel, the AGL at the time of the
  observation, the angular bin which contained the center
  of the pronghorn cluster(group), and the number of
  pronghorn in the group. The angular bins were defined by
  a combination of two window and five wing strut marks to
  define bin cutpoints for perpendicular ground distances
  of 0-65, 65-90, 90-115, 115-165 and 165-265 meters when
  the plane is 300' (91.4 meters) above ground level. The
  inner band is considered a blind region due to
  obstruction of view beneath the plane; thus th the line
  is offset 65 meters from underneath the plane.
}
\docType{data}
\references{Laake, J., R. J. Guenzel, J. L. Bengtson, P. Boveng, M. Cameron,
and M. B. Hanson. 2008.  Coping with variation in aerial survey protocol
for line-transect sampling. Wildlife Research 35:289-298.}
\keyword{datasets}
