\name{book.tee.data}
\alias{book.tee.data}
\title{Golf tee data used in chapter 6 of Advanced Distance Sampling examples...}

\description{
  Golf tee data used in chapter 6 of Advanced Distance
  Sampling examples Double platform data collected in a
  line transect survey of golf tees by 2 observers at St.
  Andrews. Field sex was actually colour of the golf tee: 0
  - green; 1 - yellow. Exposure was either low (0) or
  high(1) depending on height of tee above the ground. size
  was the number of tees in an observed cluster.
}
\docType{data}
\references{Need to add this}
\keyword{datasets}
