% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.function.r
\name{ds.function}
\alias{ds.function}
\title{Distance Sampling Functions}
\usage{
ds.function(model, newdata = NULL, obs = "All", conditional = FALSE,
  pdf = TRUE, finebr)
}
\arguments{
\item{model}{model object}

\item{newdata}{dataframe at which to compute values; if NULL uses fitting data}

\item{obs}{1 or 2 for observer 1 or 2, 3 for duplicates, "." for combined and "All" to return all of the values}

\item{conditional}{if FALSE, computes p(x) based on distance detection function and if TRUE based on mr detection function}

\item{pdf}{if FALSE, returns p(x) and if TRUE, returns p(x)*pi(x)/integral p(x)*pi(x)}

\item{finebr}{fine break values over which line is averaged}
}
\value{
List containing \item{xgrid }{grid of distance values} \item{values }{average detection fct values at the xgrid values}
}
\description{
Computes values of conditional and unconditional detection functions and
probability density functions for for line/point data for single observer or
dual observer in any of the 3 configurations (io,trial,rem).
}
\details{
Placeholder -- Not functional ----
}
\author{
Jeff Laake
}
