% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/errors.R
\name{errors}
\alias{errors}
\title{Error function}
\usage{
errors(errmsg = NULL, preamble = "Warning")
}
\arguments{
\item{errmsg}{the message to be stored/printed (optional)}

\item{preamble}{character string to paste before the message}
}
\value{
None
}
\description{
Writes error messages for various errors that can occur
}
\author{
Dave Miller
}

