\name{errors}
\alias{errors}
\title{Error function}
\usage{
  errors(errmsg = NULL, errmode = "report")
}
\arguments{
  \item{errmsg}{the message to be stored/printed
  (optional)}

  \item{errmode}{report or print errors (default report)}
}
\value{
  None
}
\description{
  Writes error messages for various errors that can occur
}
\author{
  Dave Miller
}

