% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstdint.R
\name{gstdint}
\alias{gstdint}
\title{Integral of pdf of distances}
\usage{
gstdint(x, ddfobj, index = NULL, select = NULL, width, standardize = TRUE,
  point = FALSE, stdint = TRUE, doeachint = FALSE, left = left)
}
\arguments{
\item{x}{lower, upper value for integration}

\item{ddfobj}{distance detection function specification}

\item{index}{specific data row index}

\item{select}{logical vector for selection of data values}

\item{width}{truncation width}

\item{standardize}{if \code{TRUE}, divide through by the function evaluated at 0}

\item{point}{logical to determine if point (\code{TRUE}) or line transect(\code{FALSE})}

\item{stdint}{if \code{TRUE}, scale=1 otherwise specified scale used}

\item{doeachint}{if \code{TRUE} perform integration using \code{\link{integrate}}}

\item{left}{left truncation width}
}
\value{
vector of integral values of detection function
}
\description{
Computes the integral of \code{distpdf} with scale=1 (\code{stdint=TRUE}) or specified scale (\code{stdint=FALSE}).
}
\note{
This is an internal function that is not intended to be invoked directly.
}
\author{
Jeff Laake and David L Miller
}
\keyword{utility}
