% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integratepdf.r
\name{integratepdf}
\alias{integratepdf}
\title{Numerically integrate pdf of observed distances over specified ranges}
\usage{
integratepdf(ddfobj, select, width, int.range, standardize = TRUE,
  point = FALSE, left = 0, doeachint = FALSE)
}
\arguments{
\item{ddfobj}{distance detection function specification}

\item{select}{logical vector for selection of data values}

\item{width}{truncation width}

\item{int.range}{integration range matrix; vector is converted to matrix}

\item{standardize}{logical used to decide whether to divide through by the function evaluated at 0}

\item{point}{logical to determine if point count (\code{TRUE}) or line transect (\code{FALSE})}

\item{left}{left truncation width}

\item{doeachint}{calculate each integral numerically}
}
\value{
vector of integral values - one for each observation
}
\description{
Computes integral of pdf of observed distances over x for each observation.
The method of computation depends on argument switches set and the type of
detection function.
}
\author{
Jeff Laake & Dave Miller
}
\keyword{utility}
