\docType{package}
\name{mrds-package}
\alias{mrds}
\alias{mrds-package}
\title{Mark-Recapture Distance Sampling (mrds)}
\description{
  This package implements mark-recapture distance sampling
  methods as described in D.L. Borchers, W. Zucchini and
  Fewster, R.M. (1988), "Mark-recapture models for line
  transect surveys", Biometrics 54: 1207-1220. and Laake,
  J.L. (1999) "Distance sampling with independent
  observers: Reducing bias from heterogeneity by weakening
  the conditional independence assumption." in Amstrup,
  G.W., Garner, S.C., Laake, J.L., Manly, B.F.J., McDonald,
  L.L. and Robertson, D.G. (eds) "Marine mammal survey and
  assessment methods", Balkema, Rotterdam: 137-148 and
  Borchers, D.L., Laake, J.L., Southwell, C. and Paxton,
  C.L.G. "Accommodating unmodelled heterogeneity in
  double-observer distance sampling surveys". 2006.
  Biometrics 62:372-378.)
}
\details{
  \tabular{ll}{ Package: \tab mrds \cr Type: \tab
  Package\cr Version: \tab 2.0.5\cr Date: \tab 2012-3-23\cr
  License: \tab GPL (>=2)\cr LazyLoad: \tab yes\cr }
}
\author{
  Jeff Laake <jeff.laake@noaa.gov>, David Borchers
  <dlb@mcs.st-and.ac.uk>, Len Thomas
  <len@mcs.st-and.ac.uk>, David Miller <dlm22@st-and.ac.uk>
  and Jon Bishop <jonb@mcs.st-and.ac.uk>
}
\keyword{package}

