% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrds-package.R
\docType{package}
\name{mrds-package}
\alias{mrds-package}
\alias{mrds}
\title{Mark-Recapture Distance Sampling (mrds)}
\description{
This package implements mark-recapture distance sampling
    methods as described in D.L. Borchers, W. Zucchini and Fewster,
    R.M. (1988), "Mark-recapture models for line transect surveys",
    Biometrics 54: 1207-1220. and Laake, J.L. (1999) "Distance sampling
   with independent observers: Reducing bias from heterogeneity by
    weakening the conditional independence assumption." in Amstrup,
    G.W., Garner, S.C., Laake, J.L., Manly, B.F.J., McDonald, L.L. and
    Robertson, D.G. (eds) "Marine mammal survey and assessment
    methods", Balkema, Rotterdam: 137-148 and Borchers, D.L., Laake,
    J.L., Southwell, C. and Paxton, C.L.G. "Accommodating unmodelled
    heterogeneity in double-observer distance sampling surveys". 2006.
    Biometrics 62:372-378.)
}
\details{
Further information on distance sampling methods and example code is available at \url{http://distancesampling.org/R/}.

For help with distance sampling and this package, there is a Google Group \url{https://groups.google.com/forum/#!forum/distance-sampling}.
}
\author{
Jeff Laake <jeff.laake@noaa.gov>,
        David Borchers <dlb@mcs.st-and.ac.uk>,
        Len Thomas <len@mcs.st-and.ac.uk>,
        David L. Miller <dave@ninepointeightone.net>,
        Jon Bishop <jonb@mcs.st-and.ac.uk>
}
\keyword{package}
