% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.det.R
\name{p.det}
\alias{p.det}
\title{Double-platform detection probability}
\usage{
p.det(dpformula, dplink, dppars, dpdata)
}
\arguments{
\item{dpformula}{formula for detection function}

\item{dplink}{link function ("logit","loglog","cloglog")}

\item{dppars}{parameter vector}

\item{dpdata}{double platform data}
}
\value{
vector of predicted detection probabilities
}
\description{
Computes detection probability for detection function computed from
mark-recapture data with possibly different link functions.
}
\author{
?????
}
