% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.det.tables.R
\name{print.det.tables}
\alias{print.det.tables}
\title{Print results of observer detection tables}
\usage{
\method{print}{det.tables}(x, ...)
}
\arguments{
\item{x}{result of call to ddf}

\item{\dots}{unused unspecified arguments for generic print}
}
\value{
None
}
\description{
Provides formatted output for detection tables
}
\seealso{
\code{\link{plot.det.tables}}
}
\author{
Jeff Laake
}
\keyword{utility}
