% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.ds.R
\name{print.summary.ds}
\alias{print.summary.ds}
\title{Print summary of distance detection function model object}
\usage{
\method{print}{summary.ds}(x, ...)
}
\arguments{
\item{x}{a summary of \code{ddf} model object}

\item{\dots}{unspecified and unused arguments for S3 consistency}
}
\description{
Provides a brief summary of data and fitted detection probability model
parameters, model selection criterion, and optionally abundance in the
covered (sampled) region and its standard error. What is printed depends
on the corresponding call to summary.
}
\seealso{
\code{\link{summary.ds}}
}
\author{
Jeff Laake
}
\keyword{utility}
