% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.region.dht.R
\name{survey.region.dht}
\alias{survey.region.dht}
\title{Extrapolate Horvitz-Thompson abundance estimates to entire surveyed region}
\usage{
survey.region.dht(Nhat.by.sample, samples, width, point)
}
\arguments{
\item{Nhat.by.sample}{dataframe of abundance by sample}

\item{samples}{samples table}

\item{width}{transect width}

\item{point}{if TRUE point count otherwise line transect}
}
\value{
Revised Nhat.by.sample dataframe containing estimates extrapolated
  to survey region
}
\description{
Extrapolate Horvitz-Thompson abundance estimates to entire surveyed region
}
\note{
Internal function called by \code{\link{dht}} and related functions.
}
\author{
Jeff Laake
}
\keyword{utility}
